Break `userChrome` into smaller file for import
Extract the main content of `userChrome.css` into `userChrome_sub.css`
and import into `userChrome_main.css`. `userChrome_main.css` will be
copied into `chrome` instead of `userChrome.css` and will then be
renamed to `userChrome.css` and used together with `userChrome_sub.css`.
The point of this to allow the content of `userChrome.css` be
imported and modified by extra files, while still keep it up-to-date
with upstream repo.